fetch ssn
