+++ src/main/java/com/hedvig/claims/commands/AddAutomaticPaymentCommand.kt
@@ -1,6 +1,5 @@
 package com.hedvig.claims.commands
 
-import com.hedvig.claims.web.dto.SelectedPayoutDetails
 import org.axonframework.commandhandling.TargetAggregateIdentifier
 import javax.money.MonetaryAmount
