+++ src/main/java/com/hedvig/claims/serviceIntegration/memberService/dto/Member.kt
@@ -6,6 +6,7 @@ data class Member(
     val memberId: String,
     val firstName: String? = null,
     val lastName: String? = null,
+    val ssn: String? = null,
     val birthDate: LocalDate? = null,
     val street: String? = null,
     val city: String? = null,
